Move some inline tools to shared folder.
