+++ src/article/shared/InsertCrossReferenceCommand.js
