+++ src/article/shared/InsertFootnoteCrossReferenceCommand.js
