+++ src/article/shared/InsertInlineNodeCommand.js
